packages/web_crawl_tool/pyproject.toml
[project]
name = "web-crawl-tool"
version = "0.1.0"
description = "Web crawl CLI tool for extracting readable content from web pages"
requires-python = ">=3.12"
dependencies = [
    "cli-common",
    "httpx>=0.27",
    "pydantic>=2.0",
    "readability-lxml>=0.8",
    "typer>=0.9",
]

[project.scripts]
web-crawl-tool = "web_crawl_tool.cli:app"

[tool.uv.sources]
cli-common = { workspace = true }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
mypy_path = "src"

[[tool.mypy.overrides]]
module = "cli_common.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "readability.*"
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src/web_crawl_tool"]
